+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `GenerateSubmissionDataUrl` to `FormsClient`
+- `GetMyFormsApp`, `CreateUser` and `DeleteUser` to `FormsAppClient`
 
 ## 1.17.0 (2021-03-23)